Add error correction using first world order projection (not active by default), change the Shape class into Collider class, add the new decimal type in order to easily switch between single and double precision
git-svn-id: https://reactphysics3d.googlecode.com/svn/trunk@460 92aac97c-a6ce-11dd-a772-7fcde58d38e6

+++ src/collision/ContactInfo.cpp
@@ -30,7 +30,7 @@ using namespace reactphysics3d;
 
 
 // Constructor for GJK
-ContactInfo::ContactInfo(Body* body1, Body* body2, const Vector3& normal, double penetrationDepth,
+ContactInfo::ContactInfo(Body* body1, Body* body2, const Vector3& normal, decimal penetrationDepth,
                          const Vector3& localPoint1, const Vector3& localPoint2,
                          const Transform& transform1, const Transform& transform2)
             : body1(body1), body2(body2), normal(normal), penetrationDepth(penetrationDepth),
